Initialize Stripe when the rest of the theme scripts load
This fixes an issue where Stripe was attempting to initialize before jQuery was ready because it was executing in it's place in the document order when the scripts for the theme are placed at the end of the body element so they haven't been loaded yet. Putting this script inside of the {% scripts %} tag will ensure that it's always loaded after jQuery (assuming that the theme developer has correctly implemented the `{% scripts %}` tag. Either way it's more robust than the current approach.